PCommandBot: Do not lowercase commands and channels in this class
This commit introduces a new map class that is responsible for
normalizing keys. PCommandBot no longer has to convert commands and
channels to lowercase to access the responders, synonymous responders,
or user lists.

However, nicks in the user lists are not normalized; this may be solved
in the future with a similar set class that normalizes elements.